pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "joblib>=1.4.2",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "scikit-learn>=1.6.1",
    "streamlit>=1.44.1",
    "tensorflow>=2.14.0",
]
